Skeleton.html
Here's the skeleton of the site.

It is responsive to 480px, the recommended width of an iPhone. The base
width of the wrap is 980px and scales down from there.

jQuery plugins used:
- Sidr (http://www.berriart.com/sidr/): For the left hand menu. Sorry,
it's not going to switch to right-hand while in phone mode without
rewriting it.
- Google Prettify (https://code.google.com/p/google-code-prettify/): For
the easy code colorisation/layout. Note: The <pre> tags are
tab-sensitive and needs to be left-aligned on the page, else there are
going to be large tabs on the screen.

Make any cosmetic changes that you wish. It should convert into Tumblr
okay. Everything should be commented or straight forward, and
you-break-it-you-bought-it rules are in effect.